+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Write the date in place of the "Unreleased" in the case a new version is release
 
 - Add adapters for reading back assets with the image/jpeg and
   multipart/related;type=image/jpeg mimetypes.
+- Refactor CSVAdapter to allow pd.read_csv kwargs
 
 ## v0.1.0b10 (2024-10-11)